requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
pydantic
python-dotenv
langchain
langchain-google-genai>=2.0.0
google-generativeai
requests

